// frontend/src/api/generated/models/trashedEpic.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { UserPublic } from './userPublic';

export interface TrashedEpic {
  id: number;
  team_id: number;
  name: string;
  color: string;
  ticket_count: number;
  deleted_at: string;
  deleted_by?: UserPublic | null;
  purge_at: string;
}